backend/requirements.txt
# FastAPI and server dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Security dependencies
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=3.4.8
itsdangerous>=2.0.0

# YouTube and video processing
yt-dlp>=2023.10.0
moviepy>=1.0.3

# Background tasks
celery>=5.3.0
redis>=5.0.0

# Other utilities
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.25.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0  # For TestClient

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0